requirements.txt
aiogram~=2.14
environs~=8.0.0
pandas
asyncpg~=0.28.0
requests~=2.31.0
python-slugify~=8.0.1